添加log模块线程池支持
